Rename mixer.rs to sw_mixer.rs
